feat: Initial packit setup
This commit contains the specfile for building the official package for Fedora
with a Packit setup.

Ultimately, a unified specfile is targeted for Fedora and any other rpm-based
distributions, e.g. openEuler.

And Packit(https://packit.dev/) is a tool for maintaining specfile within
upstream source. It requires a simple config file(.packit.yaml).

Log:

Signed-off-by: Robin Lee <cheeselee@fedoraproject.org>
Change-Id: I457c549bfcb77610fd6c54e183afb65d9932a291
Reviewed-on: http://gerrit.uniontech.com/c/nanjing/dde-calendar/+/9805
Reviewed-by: <mailman@uniontech.com>
Reviewed-by: kongyunzhen <kongyunzhen@uniontech.com>
Tested-by: <mailman@uniontech.com>